Removing `backup_status` and `clear` tasks
Neither of these is useful in the current workflow and really just add
to confusion about how the program should be used.

`dots clear` removed links and created `.bak` files from existing files.
At this point, it's hard for me to understand what the purpose of this
was, so instead of trying to figure out the confusing functionality, I'm
removing it. This also really doesn't have a place since the current
workflow of `add`, `resore`, `link`, and `unlink` should contain all of
the necessary functionality (mean `.bak` files don't seem necessary).

`dots backup_status` only existed to provide some insight into what was
going on with the `.bak` files, but if they are gone then this is no
longer necessary.