Adding `dots update` task for running package update scripts
This is the first step in phasing out the individual commands for
running package and configuration update scripts. These tasks should
ideally be dynamically generated based on the files that are present in
the target folder, and this task provides the first example of how to
approach accomplishing that.

For now, the existing functions are still present, to be removed at a
future point when it's clear that this new function is working properly.