Rename $DOTFILES to $DOTSPATH, default to ~/.dots and add support for .dotsrc
$DOTFILES is too generic and doesn't do a good job of describing what
the purpose of the variable is, whereas $DOTSPATH reflects the name of
the program and is arguably more conventional.

Although ~/.dotfiles is a more standard location for dotfiles, it seems
easier to keep the `dots` program agnostic by using consistent naming and,
therefore, creating a ~/.dots directory by default.

In order to make $DOTSPATH easier to customize, adding support for a
.dotsrc file. It also makes sense that a program for managing dotfiles
should itself support a dotfile.

resolves #1
resolves #2
resolves #3