Use a fancier and more explicit debug function.
Set $debug to an initial value to avoid unexpected behavior, then test
more explicitly for the value of the variable.

Include an emoji bug in output and end with a trailing horizontal rule
in order to make it more obvious where the beginning and end of the
debug statement are.