Reformat _contains() for clarity, expect array as string of elements
Use format that's consisten with rest of program, avoiding single line
statements when possible.

Use more consistent name for '_test_element' parameter. Also, avoiding
'local' keyword in favor of underscore prefix. At the moment, 'local'
isn't used elsewhere in the program, so consistency is preferred. If
'local' is reintroduced, it should probably be used throughout the
program.

Most importantly, expecting list argument to be a string of elements
should resolve some of the ambiguity around how strings are handled and
avoid errors on older versions of bash.