Set IFS to newline / tab and update command loading
Setting IFS makes it explicit what ends up acting as a terminator and
makes it easier to work with arrays and avoid errors related to spaces
in path names.

Setting IFS also reveals issues with the the implementation of command
loading, including highlighting that the `--raw` option was never
working in the intended way. Resolving these issues requires splitting
out command loading from the `commands` command and rewiriting the
function parsing.