Replace `update_submodules` with `submodules ( sync | update )`
The `update_submodules` command doesn't provide an easy way to deal with
populating submodules after pulling from a remote with submodules that
have been updated by a different clone. It also recursively updates
submodules that are nested deeper than the first level, which can be
undesirable.

In order to try to handle both cases of submodule updating, the
`update_submodules` command can be rewritten as a simpler `submodules`
command with two subcommands:

- `subdmoules sync`, which recursively updates all submodules to the
  commits specified in the parent repository or submodule,
- `submodules update`, which pulls the lastest changes from each
  root-level submodule's remote origin.