Enable argument passing for _runner() scripts
_runner() is used to run scripts in the following directories:

- bootstrap
- configure
- customize
- install

So far it's been assumed that these scripts would be unsophisticated and
simply run a series of commands. However, there do appear to be some use
cases where it would be desirable to include scripts that take
arguments. In order to facilitate the use of more sophisticated scripts,
arguments can be passed through to the script as it is called.

Currently the implementation has a couple caveats. First, the `--all`
option is used by the _runner() script, and therefore can't be easily
used in child scripts without likely encountering unexpected behavior.
Similarly, there are other options that likely won't be passed through,
such as the `--debug` option.