Use `$_COMMAND_ARGV` for parameters in `add` and enhance error checking
`$_COMMAND_ARGV` provides more logical indexing since the passed
parameters start at the '1' index.

`add` also includes a typo wherein it attempts to extract a filename
from the whole `$_arguments` array, but only the first argument should
be used.

Finally, the error checking for an existing filename doesn't account for
the length of `$_arguments`, so this is changed to check for a blank
value at index '1'.