Reorganize `$BINDLEPATH` assignment and validation.
Some functions don't need to have an assigned `$BINDLEPATH`, and
requiring the existence of a directory at that location causes the
program to exit for new users who do not have a bindle repository
created and configured yet. Therefore, exclude the relevant subcommands
from the `$BINDLEPATH` checks.

Since `$BINDLEPATH` is an application configuration variable, place its
assignment and validation in the 'Environment' section, removing the
legacy `$DOTSPATH` support in the process.