Add comment and directive to handle ShellCheck SC1090
SC1090 ("Can't follow non-constant source. Use a directive to specify
location.") is triggered due to `source` being given a non-constant
argument. Since a naked shellcheck directive could be confusing to
someone unfamiliar with shellcheck now or in the future, include
a comment describing the background reasoning behind the directive.

It's less than ideal to have a directive with comments directly in the
source code, but currently there is no easy way that I'm aware of for
excluding rules on a per-project basis.