Remove `$_CONTEXT` and support for `bin` directory linking.
This feature was originally added as a way to support the inclusion of
executables in the dotfiles repository. The functionality was
implemented by layering the handling of a second context with all of the
functionality of the `$HOME` directory linking.

This additional layer adds complexity and requires more overhead to
follow the code. There are also systems where the the `~/bin` directory
is not writable to the user. This feature is also not necessary since
`$BINDLEPATH/bin` or `$BINDLEPATH/home/.<directory>` can just be added
to `$PATH` directly.

Thic commit removes all of the contextual code and related
documentation, leaving core functionality unaltered.